Splut ui into core and conf
