+++ .gitignore
@@ -6,7 +6,7 @@ __pycache__/
 .pytest_cache
 .venv/
 enrolment/staticfiles/
-ui/staticfiles/
+*/staticfiles/
 BrowserStackLocal
 node_modules/
