+++ conf/settings.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 """
-Django settings for ui project.
+Django settings for Export Readiness project.
 
 Generated by 'django-admin startproject' using Django 1.9.6.
 
@@ -47,7 +47,6 @@ INSTALLED_APPS = [
     "django.contrib.sitemaps",
     "formtools",
     "corsheaders",
-    "ui",
     "directory_constants",
     "core",
     "article",
@@ -71,12 +70,12 @@ MIDDLEWARE_CLASSES = [
     'django.middleware.locale.LocaleMiddleware',
     'sso.middleware.SSOUserMiddleware',
     'core.middleware.NoCacheMiddlware',
-    'ui.middleware.LocaleQuerystringMiddleware',
-    'ui.middleware.PersistLocaleMiddleware',
-    'ui.middleware.ForceDefaultLocale',
+    'core.middleware.LocaleQuerystringMiddleware',
+    'core.middleware.PersistLocaleMiddleware',
+    'core.middleware.ForceDefaultLocale',
 ]
 
-ROOT_URLCONF = 'ui.urls'
+ROOT_URLCONF = 'conf.urls'
 
 TEMPLATES = [
     {
@@ -99,7 +98,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'ui.wsgi.application'
+WSGI_APPLICATION = 'conf.wsgi.application'
 
 
 # # Database
@@ -332,10 +331,6 @@ CONTACT_ZENDESK_URL = os.environ['CONTACT_ZENDESK_URL']
 CONTACT_ZENDESK_TOKEN = os.environ['CONTACT_ZENDESK_TOKEN']
 CONTACT_ZENDESK_USER = os.environ['CONTACT_ZENDESK_USER']
 
-FEATURE_CONTACT_US_ENABLED = os.getenv(
-    'FEATURE_CONTACT_US_ENABLED', 'false'
-) == 'true'
-
 LANDING_PAGE_VIDEO_URL = os.getenv(
     'LANDING_PAGE_VIDEO_URL',
     (
@@ -347,15 +342,8 @@ LANDING_PAGE_VIDEO_URL = os.getenv(
 # directory CMS
 CMS_URL = os.environ['CMS_URL']
 CMS_SIGNATURE_SECRET = os.environ['CMS_SIGNATURE_SECRET']
-FEATURE_CMS_ENABLED = os.getenv('FEATURE_CMS_ENABLED', 'false') == 'true'
-FEATURE_PERFORMANCE_DASHBOARD_ENABLED = os.getenv(
-    'FEATURE_PERFORMANCE_DASHBOARD_ENABLED', 'false') == 'true'
-
 
 # Internal CH
-FEATURE_USE_INTERNAL_CH_ENABLED = os.getenv(
-    'FEATURE_USE_INTERNAL_CH_ENABLED',
-) == 'true'
 INTERNAL_CH_BASE_URL = os.getenv('INTERNAL_CH_BASE_URL')
 INTERNAL_CH_API_KEY = os.getenv('INTERNAL_CH_API_KEY')
 
@@ -370,3 +358,17 @@ GEOLOCATION_MAXMIND_DATABASE_FILE_URL = os.getenv(
         'GeoLite2-Country.tar.gz'
     )
 )
+
+# feature flags
+FEATURE_USE_INTERNAL_CH_ENABLED = os.getenv(
+    'FEATURE_USE_INTERNAL_CH_ENABLED',
+) == 'true'
+FEATURE_CONTACT_US_ENABLED = os.getenv(
+    'FEATURE_CONTACT_US_ENABLED', 'false'
+) == 'true'
+FEATURE_CMS_ENABLED = os.getenv('FEATURE_CMS_ENABLED', 'false') == 'true'
+FEATURE_PERFORMANCE_DASHBOARD_ENABLED = os.getenv(
+    'FEATURE_PERFORMANCE_DASHBOARD_ENABLED', 'false') == 'true'
+FEATURE_SEARCH_ENGINE_INDEXING_DISABLED = os.getenv(
+    'FEATURE_SEARCH_ENGINE_INDEXING_DISABLED'
+) == 'true'
