+++ core/mixins.py
@@ -28,3 +28,15 @@ class GetCMSPageMixin:
             page=handle_cms_response(response),
             *args, **kwargs
         )
+
+
+class TranslationsMixin:
+
+    def dispatch(self, request, *args, **kwargs):
+        translation.activate(request.LANGUAGE_CODE)
+        return super().dispatch(request, *args, **kwargs)
+
+    def get_context_data(self, *args, **kwargs):
+        context = super().get_context_data(*args, **kwargs)
+        context['LANGUAGE_BIDI'] = translation.get_language_bidi()
+        return context
