+++ docker/env.test.json
@@ -58,7 +58,8 @@
       "FEATURE_USE_INTERNAL_CH_ENABLED",
       "INTERNAL_CH_BASE_URL",
       "INTERNAL_CH_API_KEY",
-      "GEOLOCATION_MAXMIND_DATABASE_FILE_URL"
+      "GEOLOCATION_MAXMIND_DATABASE_FILE_URL",
+      "FEATURE_SEARCH_ENGINE_INDEXING_DISABLED"
     ]
   }
 }
