--- ui/views.py
@@ -1,13 +1,0 @@
-from django.utils import translation
-
-
-class TranslationsMixin:
-
-    def dispatch(self, request, *args, **kwargs):
-        translation.activate(request.LANGUAGE_CODE)
-        return super().dispatch(request, *args, **kwargs)
-
-    def get_context_data(self, *args, **kwargs):
-        context = super().get_context_data(*args, **kwargs)
-        context['LANGUAGE_BIDI'] = translation.get_language_bidi()
-        return context
